requirements.txt
click==8.0.4
datasets==2.2.2
einops==0.4.1
ftfy==6.1.1
gpustat==0.6.0
ipdb==0.13.9
matplotlib==3.5.1
Markdown==3.3.6
numpy==1.22.3
opencv-python==4.5.5.64
pandas==1.4.1
Pillow==9.0.1
protobuf==3.19.4
pyarrow==7.0.0
pytorch-lightning==1.5.10
requests==2.27.1
tqdm==4.63.0
sacred==0.8.2
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
tensorboard==2.8.0
timm==0.5.4
tokenizers==0.11.6
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchmetrics==0.7.2
torchvision==0.12.0+cu113
transformers==4.17.0
urllib3==1.26.8
wandb==0.12.11